UI/BE: Refactor (extract semanctic getters from technical client) and wire up routes
