+++ HoloRepositoryUI-Server/src/api/routes/patients/patients.service.ts
@@ -1,7 +1,8 @@
 import logger from "../../../common/logger";
 import { IPatient } from "../../../../../HoloRepositoryUI-Types";
+import { getPatient } from "../../../common/data.service";
 
-import samplePatientsAll from "../../../__tests__/samples/samplePatientsAll.json";
+import samplePatientsAll from "../../../__tests__/samples/internal/samplePatientsAll.json";
 
 const _samplePatients = samplePatientsAll as IPatient[];
 
@@ -13,10 +14,7 @@ export class PatientsService {
 
   public getById(pid: string): Promise<IPatient> {
     logger.info(`GET Patient by id '${pid}'`);
-    const patient = samplePatientsAll.find(patient => patient.pid === pid);
-    if (patient) {
-      return Promise.resolve(patient as IPatient);
-    }
+    return getPatient(pid);
   }
 }
