+++ HoloRepositoryUI-Server/src/api/routes/practitioners/practitioners.service.ts
@@ -1,12 +1,11 @@
 import logger from "../../../common/logger";
 import { IPractitioner } from "../../../../../HoloRepositoryUI-Types";
-
-import samplePractitioner from "../../../__tests__/samples/samplePractitioner.json";
+import { getPractitioner } from "../../../common/data.service";
 
 export class PractitionersService {
   public getById(pid: string): Promise<IPractitioner> {
     logger.info(`GET Practitioner by id '${pid}'`);
-    return Promise.resolve(samplePractitioner as IPractitioner);
+    return getPractitioner(pid);
   }
 }
